Update to 1.21

+++ src/main/java/net/barch/brosalch/Entities/BoltOfIceEntity.java
@@ -56,7 +56,7 @@ public class BoltOfIceEntity extends ThrownItemEntity {
     }
 
     @Override
-    protected float getGravity() {
+    protected double getGravity() {
         return 0;
     }
